JBAI-4393 [ndarray] Functional interface to streamline parallelization block handling (avoid Integer boxing operations).

+++ ndarray/ndarray-core/src/jvmMain/kotlin/io/kinference/ndarray/extensions/constants/BooleanConstants.kt
@@ -3,5 +3,5 @@ package io.kinference.ndarray.extensions.constants
 object BooleanConstants {
     const val ZERO = false
     const val ONE = true
-    const val SIZE_BYTES = 1.toLong()
+    const val SIZE_BYTES = 1L
 }
